requirements.txt
py-cord
pillow
requests
